feat: loading optimize & adjust some css

+++ src/lib/storage.ts
@@ -1,20 +1,19 @@
+import { LocalStorageKey } from '@/enums/index.enum';
+import { getLocalStorage, setLocalStorage } from '@/lib/utils';
 import { z } from 'zod';
 
 // Define the current version of localStorage schema
 export const STORAGE_VERSION = '0.3.9';
 
-// Define the version key in localStorage
-const VERSION_KEY = '_storage_version';
-
 // Define the schema for version check
-const versionSchema = z.string();
+const versionSchema = z.string().regex(/^\d+\.\d+\.\d+$/);
 
 /**
  * Check and migrate localStorage data based on version
  * This function should be called when the app initializes
  */
 export const checkAndMigrateStorage = (): boolean => {
-  const currentVersion = localStorage.getItem(VERSION_KEY);
+  const currentVersion = getLocalStorage(LocalStorageKey.VERSION_KEY, versionSchema);
 
   // If no version exists or version doesn't match, clear all storage
   if (
@@ -22,15 +21,19 @@ export const checkAndMigrateStorage = (): boolean => {
     !versionSchema.safeParse(currentVersion).success ||
     currentVersion !== STORAGE_VERSION
   ) {
-    // eslint-disable-next-line no-console
-    console.log(
+    if (currentVersion === undefined) {
+      setLocalStorage(LocalStorageKey.VERSION_KEY, STORAGE_VERSION, versionSchema);
+      return false
+    }
+
+    console.warn(
       `[STORAGE] currentVersion=${currentVersion} is outdated (expectedVersion=${STORAGE_VERSION}), need to clear all storage`
     );
     // Clear all localStorage data
     localStorage.clear();
 
     // Set the new version
-    localStorage.setItem(VERSION_KEY, STORAGE_VERSION);
+    setLocalStorage(LocalStorageKey.VERSION_KEY, STORAGE_VERSION, versionSchema);
 
     // You can add specific migration logic here if needed
     // For example:
